Topic Redefinition Proposal
The following proposal provides an alternative way to simply declare topics
that can be easily easily understood both by reading the proto and
programmatically through the generated code.  the change renames UResource
to UTopic for clarity, and creates a CustomOption to declare the topics in the
service section of the proto. Example use of the topics can be found in a
subsequent uprotocol-java code change shortly.