Urgent fix in uTwin package Definition (#114)
uTwin was changed from v1 to v2 because we were going to call the original uTwin API that used cloudevents as V1 however the package name was not fixed in the proto causing build issues.  This fix addresses that problem as well as a minor document change in uprotocol_options.proto for the topic ID number.